[Impeller] Move instanced-attributes test to the golden harness
Move CanRenderInstancedWithVertexAttributes from renderer_unittests.cc
into renderer_golden_unittests.cc so the instanced draw is captured as a
Skia Gold image. Reduce the pipeline to the golden render target
(single-sampled, no depth/stencil) and reset the host buffer per pass.
Drop the now-unused instanced_attributes includes from the playground
test.
